Make newcharacter do what it says
Several bug fixes; it pretty much was completely broken before. Fixed
permissions so there actually is one, default true. Added another
argument, [target] which requires gildorym.command.newcharacter.other.
Using this command makes a note in the generation method of the
character. There is no current limit on the number of characters (to be
fixed later). This appears to be doing everything correctly.

This makes the next step, swapping characters / looking at a list of
characters very easy!